Refactor code chnages
