--- src/components/Pages/Continent.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Continent() {
-  return <div>Continent</div>;
-}
