--- src/components/Pages/Country.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Country() {
-  return <div>Country</div>;
-}
