--- src/components/Pages/Home.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Home() {
-  return <div>Home</div>;
-}
