+++ src/redux/getContinent.js
