+++ src/redux/reducer.js
@@ -1,5 +1,5 @@
 import { combineReducers } from '@reduxjs/toolkit';
-import reducers from './combineReducers';
+import reducers from './combineReducer';
 
 export default combineReducers({
   entities: reducers,
